Jjama3

Documentation for Jjama3.

Jjama3.assistant_prompt

Jjama3.format_llama32_instruction_prompt

Jjama3.generate

Jjama3.llama3_tokenizer

Jjama3.load_llama3_from_safetensors

Jjama3.assistant_prompt — Method

generate(model, prompt, max_new_tokens=100, encoder_for_printing=tkn)

Format a prompt for use with Llama3.2's instruction format, with a simple "You are a helpful assistant" system prompt.

tkn = llama3_tokenizer()
prompt = assistant_prompt("What is the capital of France?", tkn)
generate(model, prompt, max_new_tokens=100, encoder_for_printing=tkn)

Jjama3.format_llama32_instruction_prompt — Method

generate(model, prompt, max_new_tokens=100, encoder_for_printing=tkn)

Format a prompt for use with Llama3.2's instruction format, injecting the system and user roles.

tkn = llama3_tokenizer()
prompt = format_llama32_instruction_prompt("\nYou are a helpful assistant\n", "What is the capital of France?", tkn)
generate(model, prompt, max_new_tokens=100, encoder_for_printing=tkn)

Jjama3.generate — Method

generate(model, initial_tokens; max_new_tokens=100, sampler=top_pk_sampler(p=0.5f0, k=5), encoder_for_printing=tkn, end_token=128010)

Takes an initial sequence of tokens, and generates new tokens one at a time until the end token is sampled. Uses a KV cache. No batch dim for now. Runs on CPU by default. If the model is on the GPU (assuming Flux.jl, eg. model = gpu(model)), then pass device = gpu to generate to run on the GPU.

tkn = llama3_tokenizer()
generate(model, initial_tokens; max_new_tokens=100, sampler=top_pk_sampler(p=0.5f0, k=5), encoder_for_printing=tkn, end_token=128010)

Jjama3.llama3_tokenizer — Method

tkn = llama3_tokenizer()

Load the tokenizer for Llama3. This seems to work, but I have not checked if there are some different edge-cases, or missing tokens relative to the original tokenizer (besides the special tokens we hackily include).

tkn = llama3_tokenizer()
tkn.encode("What is the capital of France?")
tkn.decode([10, 2, 5, 99])

Jjama3.load_llama3_from_safetensors — Method

model = load_llama3_from_safetensors(model_weight_paths, config)

Load a Llama3 model from a set of Huggingface safetensors files, and the config.json file. Important note: Huggingface uses a different RoPE convention than other implementations, so if you're loading weights from a different source, you might get very poor model performance.

using JSON3
config = JSON3.read(read("Llama3_2_1B_instruct/config.json", String))
model_weight_paths = ["Llama3_2_1B_instruct/model.safetensors"] #Can be an array of paths if the model is split across multiple files
model = load_llama3_from_safetensors(model_weight_paths, config)